refactor(qua_lexer): make IntLiteral u32

+++ qua_lexer/src/lib.rs
@@ -5,7 +5,7 @@ pub enum Token {
     Identifier(String), // factorial, for, int
 
     // Literals
-    IntLiteral(u16),       // 21
+    IntLiteral(u32),       // 21
     FloatLiteral(f32),     // 3.14
     StringLiteral(String), // "Qua!"
